Remove RepositoryBackend trait
The repository format depends on the implementation of multiple modules,
so having a trait only for the Repository does not add any advantage.

If the time comes when we need to support a new repository format, we will
need to find something that works for the repo, index, pack, keys, etc...
