+++ src/commands/cmd_cat.rs
@@ -21,8 +21,8 @@ use clap::Args;
 
 use crate::backend::new_backend_with_prompt;
 use crate::global::{ID, ID_LENGTH};
+use crate::repository::repo::{RepoConfig, Repository};
 use crate::repository::tree::Tree;
-use crate::repository::{self, RepoConfig};
 use crate::utils::{self, size};
 
 use super::GlobalArgs;
@@ -55,7 +55,7 @@ pub fn run(global_args: &GlobalArgs, args: &CmdArgs) -> Result<()> {
     let config = RepoConfig {
         pack_size: (global_args.pack_size_mib * size::MiB as f32) as u64,
     };
-    let (repo, _) = repository::try_open(pass, global_args.key.as_ref(), backend, config)?;
+    let (repo, _) = Repository::try_open(pass, global_args.key.as_ref(), backend, config)?;
 
     match &args.object {
         Object::Manifest => {
